[v2.2.x] prov/efa: Move RDMA check functions to efa_rdm_ep.h
Necessary to avoid circular dependencies in the next commit that moves
the peer pool to the endpoint instead of the AV

(cherry picked from commit 81cff3705e428bae778c23f7ffa88a657f694bb9)
